Revert "Title check added for window reuse in Windows.GetBySpec." Refs: #34.
This reverts commit a8269d3338fc6e0c6e94015b880d004f858049ba.

The title check did not work properly together with the Info->Source command
because the window is searched first without a title and later it is searched again with
a title that is not the real window title but the file name.
So the window is found for reuse in the first search but not in the second search resulting
in an inconsistency in the window management.

Signed-off-by: Josef Templ <josef.templ@gmail.com>