bug fix in importing RTF texts with tables. Refs: #86. As proposed by Helmut Zinn, writing a CR at the end of a table row is already done by Paragraph() and has been removed.
In addition, openBrackets has been removed because it turned out not to be required.
The current context (con) can be used for detecting if a closing bracket belongs to a
hyperlink. The same principle is now also applied for detecting if a table row is processed.

In addition, a subtle bug has been fixed with the the detection of the end of an RTF import. When importing via the the Clipboard, the MemReader does not detect the end of the RTF data correctly but delivers undefined characters until it hits a 0X. Therefore, ParseRichText now terminates when the outermost bracket is closed.

Signed-off-by: Josef Templ <josef.templ@gmail.com>